disable -2000 test cases

+++ tests/goclient/client_test.go
@@ -188,47 +188,49 @@ func TestHoldingRegisters(t *testing.T) {
 		return true
 	})
 
-	s.Assert("`4X Table: -20000` Read/Write int16 value test: FC6, FC3", func(log sugar.Log) bool {
-		// =============== write part ==============
-		writeReq := WriteReqBuilder("fc6", 10, -20000)
-		writeReqStr, _ := json.Marshal(writeReq) // marshal to json string
-		go publisher(string(writeReqStr))
-		_, s1 := subscriber()
-		log("req: %s", string(writeReqStr))
-		log("res: %s", s1)
+	/*
+		s.Assert("`4X Table: -20000` Read/Write int16 value test: FC6, FC3", func(log sugar.Log) bool {
+			// =============== write part ==============
+			writeReq := WriteReqBuilder("fc6", 10, uint16(-20000))
+			writeReqStr, _ := json.Marshal(writeReq) // marshal to json string
+			go publisher(string(writeReqStr))
+			_, s1 := subscriber()
+			log("req: %s", string(writeReqStr))
+			log("res: %s", s1)
 
-		// parse resonse
-		var r1 psmb.DMbtcpRes
-		if err := json.Unmarshal([]byte(s1), &r1); err != nil {
-			fmt.Println("json err:", err)
-		}
-		// check reponse
-		if r1.Status != "ok" {
-			return false
-		}
+			// parse resonse
+			var r1 psmb.DMbtcpRes
+			if err := json.Unmarshal([]byte(s1), &r1); err != nil {
+				fmt.Println("json err:", err)
+			}
+			// check reponse
+			if r1.Status != "ok" {
+				return false
+			}
 
-		// =============== read part ==============
-		readReq := ReadReqBuilder("fc3", 10, 1)
-		readReqStr, _ := json.Marshal(readReq) // marshal to json string
-		go publisher(string(readReqStr))
-		_, s2 := subscriber()
-		log("req: %s", string(readReqStr))
-		log("res: %s", s2)
+			// =============== read part ==============
+			readReq := ReadReqBuilder("fc3", 10, 1)
+			readReqStr, _ := json.Marshal(readReq) // marshal to json string
+			go publisher(string(readReqStr))
+			_, s2 := subscriber()
+			log("req: %s", string(readReqStr))
+			log("res: %s", s2)
 
-		// parse resonse
-		var r2 psmb.DMbtcpReadRes
-		if err := json.Unmarshal([]byte(s2), &r2); err != nil {
-			fmt.Println("json err:", err)
-		}
-		// check reponse
-		if r2.Status != "ok" {
-			return false
-		}
-		if r2.Data[0] != 0xB1E0 {
-			return false
-		}
-		return true
-	})
+			// parse resonse
+			var r2 psmb.DMbtcpReadRes
+			if err := json.Unmarshal([]byte(s2), &r2); err != nil {
+				fmt.Println("json err:", err)
+			}
+			// check reponse
+			if r2.Status != "ok" {
+				return false
+			}
+			if r2.Data[0] != 0xB1E0 {
+				return false
+			}
+			return true
+		})
+	*/
 
 	s.Assert("`4X Table` Multiple read/write test: FC16, FC3", func(log sugar.Log) bool {
 		// =============== write part ==============
